Changed class LoginPresenter to LoginViewModel.kt
