+++ app/src/main/java/com/svetlanakuro/mvp_mvvm_patterns/ui/login/LoginContract.kt
@@ -19,6 +19,8 @@ class LoginContract {
         val shouldShowProgress: Publisher<Boolean>
         val isSuccess: Publisher<Boolean>
         val errorText: Publisher<String>
+        val addAccountSuccess: Publisher<Boolean>
+        val resetPasswordSuccess: Publisher<String>
 
         fun onSignIn(login: String, password: String)
         fun onSignUp(login: String, password: String)
